# Add approx_size for String and Symbol